+++ modules/iam-group-with-policies/variables.tf
@@ -10,6 +10,12 @@ variable "name" {
   default     = ""
 }
 
+variable "path" {
+  description = "Desired path for the IAM group"
+  type        = string
+  default     = "/"
+}
+
 variable "group_users" {
   description = "List of IAM users to have in an IAM group which can assume the role"
   type        = list(string)
@@ -28,7 +34,7 @@ variable "custom_group_policies" {
   default     = []
 }
 
-variable "enable_mfa_enforcment" {
+variable "enable_mfa_enforcement" {
   description = "Determines whether permissions are added to the policy which requires the groups IAM users to use MFA"
   type        = bool
   default     = true